# docs/services/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: CORTX Gateway
  version: 0.1.0
paths:
...
  /propverify/{path}:
    post:
      tags:
      - PropVerify
      - PropVerify Proxy
      summary: Propverify Proxy All
      description: Proxy all requests to PropVerify service
      operationId: propverify_proxy_all_propverify__path__post
      parameters:
      - name: path
        in: path
        required: true
        schema:
          type: string
          title: Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - PropVerify
      - PropVerify Proxy
      summary: Propverify Proxy All
      description: Proxy all requests to PropVerify service
      operationId: propverify_proxy_all_propverify__path__post
      parameters:
      - name: path
        in: path
        required: true
        schema:
          type: string
          title: Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - PropVerify
      - PropVerify Proxy
      summary: Propverify Proxy All
      description: Proxy all requests to PropVerify service
      operationId: propverify_proxy_all_propverify__path__post
      parameters:
      - name: path
        in: path
        required: true
        schema:
          type: string
          title: Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - PropVerify
      - PropVerify Proxy
      summary: Propverify Proxy All
      description: Proxy all requests to PropVerify service
      operationId: propverify_proxy_all_propverify__path__post
      parameters:
      - name: path
        in: path
        required: true
        schema:
          type: string
          title: Path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError